Add Header component and remove unnecessary graphql code
