+++ src/pages/work.js
@@ -1,15 +1,10 @@
 import React from "react"
-import { graphql } from "gatsby"
-
 import Layout from "../components/layout"
 
 class Work extends React.Component {
   render() {
-    const { data } = this.props
-    const siteTitle = data.site.siteMetadata.title
-
     return (
-      <Layout siteTitle={siteTitle}>
+      <Layout>
         Work page
       </Layout>
     )
@@ -17,13 +12,3 @@ class Work extends React.Component {
 }
 
 export default Work
-
-export const pageQuery = graphql`
-  query {
-    site {
-      siteMetadata {
-        title
-      }
-    }
-  }
-`
